+++ src/components/Modal/ReservationInfoModal/Common/ReservationInfo.tsx
