+++ src/components/Modal/ReservationModal/ReservationModal.tsx
@@ -2,10 +2,10 @@ import CloseIcon from "@/assets/svgs/btnXbig.svg";
 import Button from "@/src/components/Button/Button";
 import Calendar from "@/src/components/Calendar/Calendar";
 import PopupModal from "@/src/components/Modal/PopupModal";
-import ParticipantCount from "@/src/components/Modal/ReservationModal/common/ParticipantCount";
-import Price from "@/src/components/Modal/ReservationModal/common/Price";
-import TimeSelector from "@/src/components/Modal/ReservationModal/common/TimeSelector";
-import TotalPrice from "@/src/components/Modal/ReservationModal/common/TotalPrice";
+import ParticipantCount from "@/src/components/Modal/ReservationModal/Common/ParticipantCount";
+import Price from "@/src/components/Modal/ReservationModal/Common/Price";
+import TimeSelector from "@/src/components/Modal/ReservationModal/Common/TimeSelector";
+import TotalPrice from "@/src/components/Modal/ReservationModal/Common/TotalPrice";
 import { usePostActivityReservation } from "@/src/queries/useActivities";
 import { useCalendar } from "@/src/stores/calendarStore";
 import useModalStore from "@/src/stores/modalStore";
